.sixth/skills/requirements.txt
python-telegram-bot==20.7
groq==0.4.2
gspread==5.12.0
google-auth==2.23.4
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
aiofiles==23.2.1